move mcp marketplace from client provider to its own service file
